captcha-train/pyproject.toml
[project]
name = "captcha-train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "onnx>=1.19.1",
    "onnxscript>=0.5.6",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
